move task page to pages folder and connect to backend
