+++ package.json
@@ -33,6 +33,7 @@
     "react": "^16.8.6",
     "react-adopt": "^0.6.0",
     "react-apollo": "^2.5.6",
+    "react-apollo-hooks": "^0.4.5",
     "react-dom": "^16.8.6",
     "react-konva": "^16.8.7-3",
     "react-modal": "^3.8.1",
