+++ src/pages/TaskPage/AnnotationCanvas/2DImage.js
