+++ src/pages/TaskPage/AnnotationCanvas/Canvas.js
