+++ src/pages/TaskPage/AnnotationCanvas/index.js
