+++ src/pages/TaskPage/AnnotationCanvas/styles/Canvas.css
